update chef-server-stats
Summary: This should fix the outstanding issue #2 as will as including
the provided fix for issue #1. It also supports newer versions of the
chef server both private chef and open source chef with better
auto-detection. It fully gives up on being run anywhere but locally to
the chef server.

Test Plan: Running in production for a while now.