use a more sensible value in example code
Setting vm.swappiness=70 on a server by copying it from the README
could provide surprising results. The default in the kernel is 60.

A value of 70 instructs the kernel to be more aggressive about
swapping out application data. At a minimum, this will cause needless
IO. In the worst case can render a previously slow system totally
unresponsive.

As the new text in the README says, setting it to 1 is often what
server operators want: the kernel should not swap unless it really
has to.