fix chefctl duplicate logs
Summary:
chef 13 changed the [default output to be the formatter, not the logger](https://github.com/chef/chef/blob/chef-13/RELEASE_NOTES.md#the-logger-and-formatter-settings-are-more-predictable) as follows:
> The logger and formatter settings are more predictable
> The default now is the formatter. There is no more automatic switching to the logger when logging or when output is sent to a pipe. The logger needs to be specifically requested with --force-logger or it will not show up.
> The --force-formatter option does still exist, although it will probably be deprecated in the future.
> If your logfiles switch to the formatter, you need to include --force-logger for your daemonized runs.
> Redirecting output to a file with chef-client > /tmp/chef.out now captures the same output as invoking it directly on the command line with no redirection.

Adding in the `--force-logger` so we enforce using the logger.

We also discovered a bug in chef 13 that causes duplicate log lines when the logfile is set to STDOUT but STDOUT isn't associated to a tty. Forcing chefctl to use `--logfile=/dev/null` seems to mitigate the issue while we work with upstream to figure out the underlying cause. Tracking this in T28228094, with https://github.com/chef/chef/issues/7184 and https://getchef.zendesk.com/hc/en-us/requests/18798

Reviewed By: jaymzh

Differential Revision: D7754969

fbshipit-source-id: e343c389bebff82276fd099992205602a8eae512