fix chef.first.out bug
Summary:
We saw an increasing number of hosts with no chef.first.out file. The
heuristic was broken that we used to determine if we needed to create the file.
When we started logging chefctl output to the chef log, we get output in chef logs
from blocked chefctl runs. That means if the first run takes longer than 15 minutes,
systemd queues another chefctl run, which immediately writes to a new chef.*.out
file. So we need to change our heuristic to see if the current log file is the
*oldest* log file.

Reviewed By: zfjagann

Differential Revision: D8029791

fbshipit-source-id: 82aea638a3777ed4700cb503e2a094495535b72f