Soloctl with -i shouldn't reap non-soloctl chef runs
Summary:
Right now, when chefctl is passed a `-i` argument, it will walk a process tree and find all `chefctl` processes and evaluate if that process should be reaped. This does not distinguish between chefctl process types, namely the difference between `--program chefctl` and `--program soloctl`. This results in technically different chef runs, with different code, reaping each other.

In this diff, I'm changing chefctl so it will compare processes with the same the chefctl process name (via the `program_name`), not all chefctl processes.

Reviewed By: jaymzh

Differential Revision: D9756816

fbshipit-source-id: 4b1ac7a0882d55cca1cf29b6fdca077ce5999bde