Fix stdout logging from chefctl (#75)
Summary:
This PR fixes two issues:

1. Chefctl assumes the logger from `Mixlib::Log::Logger` and `Logger` are
the same, but they are not. One's message is a has and ones is a string.
2. Mixlib::Log doesn't know about programname in the way Logger does,
so console messages were also missing that and had an awkward space.
This fills in a sane default.

With this PR messages go from:
```
[2019-07-27 00:01:11 +0000] INFO : {:msg=>"Updating chef repo"}
```
to
```
[2019-07-27 00:01:25 +0000] INFO chefctl: Updating chef repo
```

Signed-off-by: Phil Dibowitz <phil@ipom.com>
Pull Request resolved: https://github.com/facebook/chef-utils/pull/75

Differential Revision: D16552837

Pulled By: NaomiReeves

fbshipit-source-id: d8b44b2935a13921c291c0f6964fe2ee0274d8e2