Filter sush checking number of running chefctl
Summary:

chefctl checks how many chefctl processes are running/queued using ps and if there are more than 5 it fails with the "Several chef runs already queued. Not queueing any more.". If one started chefctl via sush on remote machines from the devvm and number of these chefctl procs is more than 5, chef on the devvm will fail, while these chefctl commands really run on the remote machines and not on the devvm.

Add sush to the black list.

Reviewed By: joshuamiller01

Differential Revision: D16948879

fbshipit-source-id: ad833ce5d40f494319a623584beebeb9173788b0