Improving lock file handling on Windows for chef solo
Summary:
When a file lock is held by/for chef solo, the experience on Windows was to fail with an error message that did not provide meaningful information. In addition, the user would expect that the system would retry/wait for the file lock to become available.

This chanage ensures that soloctl correctly handles the file lock, both retrying the lock operation until the timeout expires, and giving an informative error message.

Reviewed By: NaomiReeves, davide125

Differential Revision: D25108951

fbshipit-source-id: 995c544090a3a340b7f7ac38cc47071a08c2030e