Update chef-server-stats to aggregate stats across orgs
Summary:

Currently chef-server-stats reports a chef server statistics, some of
which span chef organizations, some of which don't. To make the results
more consistent, add a new option that tells chef-server-stats to
aggregate all stats across organizations.

Since chef-server-stats uses knife as its interpreter, we'll use the
knife.rb config file to store a new option to control this behavior.

We'll enable this behavior via a new chef-server-stats specific option
called 'chef_server_stats_pivotal_key'. To get a list of organizations,
chef-server-stats needs to authenticate as the pivotal user (using the
pivotal key), so if this key is available we'll query and aggregate
results across organizations.

Reviewed By: NaomiReeves

Differential Revision: D26552854

fbshipit-source-id: fa332cbf4b71ef2a3c62ecddc18fbfba8e634c72