Remove chefctl lock race condition
Summary:
There's a race in the lock cleanup and locking for chefctl.

The problem is that we remove the lock file as part of the cleanup: if run_1 is finishing up and run_2 and run_3 are waiting to start then the following can happen:

- run_1 releases lock
- run_2 grabs lock and starts running
- run_1 unlinks lock file
- run_3 starts and sees no lock file so creates one and locks it

And now we have 2 parallel runs.

By not removing the lock file run_3 is blocked until run_2 releases the lock as expected.

Reviewed By: gogsbread

Differential Revision: D35928121

fbshipit-source-id: ec6f03d9038bc16be5acd3cc3649e3487804834a