chefctl: pass rerun count down to chef-client via env var
Summary:
In order to start properly handling chef 401s, we need to change chefctl to
pass whether chefctl is rerunning the chef client.  Then the chef handlers can
use this information to tell if a rekey event is the normal rekey (and not
report an exception to scuba), or if it has
happened multiple times (in which case an exception should still be reported).

Differential Revision: D63587866

fbshipit-source-id: 9602fbb923c8008e2617d88ecc09bd512bf341f9