chefctl: fix a hard-coded path
This adds a configuration for `chef_config_dir` where we can look for
the prod cert link rather than assuming `/etc/chef` - so that those
using cinc can use chefctl more easily.

The default is, of course `/etc/chef` so nothing will change for anyone
not specifying this config.

Also: gem file update is to fix tests - rubocopy fails to run without
base64 installed:

```
/opt/hostedtoolcache/Ruby/3.4.1/x64/lib/ruby/3.4.0/bundled_gems.rb:82:in 'Kernel.require': cannot load such file -- base64 (LoadError)
```

Side note, what I really wanted to do was create a `client_distro`, and
then build the defaults based on _that_, but to allow people to still be
able to overwrite _that_ requires significant refactoring, using the new
`default(..)` syntax for mixlib-config and then creating a method for
most of the configs, and that seemed like too much change for something
so small, so I did not.

Signed-off-by: Phil Dibowitz <phil@ipom.com>