chefctl introduce --priority flag
Summary:
Despite chef (theoretically) running at Idle priority on windows, it doesn't really do that.

I found that the chef client, which seems to be responsible for most of my ruby.exe resource usage, is a subprocess spawned here:

https://www.internalfb.com/code/opsfiles/[dfde5280d732b0f7f750f09789626073e8c787ca]/scripts/chef/chefctl.rb?lines=1096-1109

For some reason, it is not inheriting the Idle priority of the parent process, despite this being typical for window applications. Per microsoft docs: https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-createprocessa?redirectedfrom=MSDN

>  If none of the priority class flags is specified, the priority class defaults to NORMAL_PRIORITY_CLASS unless the priority class of the creating process is IDLE_PRIORITY_CLASS or BELOW_NORMAL_PRIORITY_CLASS. In this case, the child process receives the default priority class of the calling process.

In the test plan, I show that many (maybe all?) of the subprocesses spawned by chefctl are now inheriting the idle priority on windows when the new flag is used.

 ---

Differential Revision: D71943120

fbshipit-source-id: 3f722c4170bc66e958d7fe1d7c36dad20d924fda